chore: breakup two largest tests (userProfile and fluxQueryBuilder) in half.
