+++ cypress/e2e/shared/scriptQueryBuilder.flux.test.ts
@@ -8,25 +8,11 @@ const DEFAULT_FLUX_EDITOR_TEXT =
 const DELAY_FOR_LAZY_LOAD_EDITOR = 30000
 const DELAY_FOR_LSP_SERVER_BOOTUP = 7000
 
-const DELAY_FOR_FILE_DOWNLOAD = 5000
-
 describe('Script Builder', () => {
   const writeData: string[] = []
   for (let i = 0; i < 30; i++) {
     writeData.push(`ndbc,air_temp_degc=70_degrees station_id_${i}=${i}`)
     writeData.push(`ndbc2,air_temp_degc=70_degrees station_id_${i}=${i}`)
-  }
-  const writeDataMoar: Array<string[]> = [[], [], [], []]
-  const numOfUniqueColumns = 198 // 200 - 2
-  for (let i = 0; i < numOfUniqueColumns; i++) {
-    writeDataMoar[0]?.push(
-      `ndbc_big,air_temp_degc=70_degrees station_id_A${i}=${i}`
-    )
-    writeDataMoar[1]?.push(
-      `ndbc_big,air_temp_degc=70_degrees station_id_A${i}=${
-        i + numOfUniqueColumns
-      }`
-    )
   }
 
   const bucketName = 'defbuck'
@@ -90,7 +76,7 @@ describe('Script Builder', () => {
       `fn: (r) => r._measurement == "${measurement}"`
     )
     cy.getByTestID('flux-editor').within(() => {
-      cy.get('.composition-sync--on').should('have.length', 3) // three lines
+      cy.get('.composition-sync--on').should('have.length.gte', 3) // three lines
     })
   }
 
@@ -152,230 +138,15 @@ describe('Script Builder', () => {
         return cy.get('@org').then(({id, name}: Organization) => {
           cy.log('add mock data')
           cy.createBucket(id, name, 'defbuck2')
-          cy.createBucket(id, name, 'defbuck3') // 0 tables, 0 rows
-          cy.createBucket(id, name, 'defbuck4') // load certain amts of data, per measurement
           cy.writeData(writeData, 'defbuck')
           cy.writeData(writeData, 'defbuck2')
-          cy.writeData(
-            [`ndbc_1table,air_temp_degc=70_degrees station_id=1`],
-            'defbuck4'
-          )
-          writeDataMoar.forEach(data => cy.writeData(data, 'defbuck4'))
         })
       })
     })
   })
 
-  describe('Results display', () => {
-    let route: string
-
-    beforeEach(() => {
-      loginWithFlags({
-        schemaComposition: true,
-        newDataExplorer: true,
-        saveAsScript: true,
-      }).then(() => {
-        cy.get('@org').then(({id: orgID}: Organization) => {
-          route = `/orgs/${orgID}/data-explorer`
-          cy.intercept(
-            'POST',
-            `/api/v2/query?${new URLSearchParams({orgID})}`,
-            req => {
-              const {extern} = req.body
-              if (
-                extern?.body[0]?.location?.source ==
-                `option v =  {  timeRangeStart: -1h,\n  timeRangeStop: now()}`
-              ) {
-                req.alias = 'query -1h'
-              } else if (
-                extern?.body[0]?.location?.source ==
-                `option v =  {  timeRangeStart: -15m,\n  timeRangeStop: now()}`
-              ) {
-                req.alias = 'query -15m'
-              }
-            }
-          )
-        })
-
-        clearSession()
-        cy.getByTestID('flux-sync--toggle').should('have.class', 'active')
-        cy.getByTestID('flux-editor', {timeout: DELAY_FOR_LAZY_LOAD_EDITOR})
-      })
-    })
-
-    it('will allow querying of different data ranges', () => {
-      selectSchema()
-      confirmSchemaComposition()
-
-      cy.log('query with default date range of -1h')
-      cy.getByTestID('time-machine-submit-button').should('exist').click()
-      cy.wait('@query -1h')
-
-      cy.log('query date range can be adjusted')
-      cy.getByTestID('timerange-dropdown').within(() => {
-        cy.getByTestID('dropdown--button').should('exist')
-        cy.getByTestID('dropdown--button').clickAttached()
-      })
-      cy.getByTestID('dropdown-item-past15m').should('exist').click()
-      cy.getByTestID('time-machine-submit-button').should('exist').click()
-      cy.wait('@query -15m')
-    })
-
-    describe('data completeness', () => {
-      const downloadsDirectory = Cypress.config('downloadsFolder')
-      const browser = Cypress.config('browser')
-
-      const validateCsv = (csv: string, tableCnt: number) => {
-        cy.wrap(csv)
-          .then(doc => doc.trim().split('\n'))
-          .then(list => {
-            expect(list[list.length - 1]).contains(`,${tableCnt - 1},`)
-          })
-      }
-
-      const runTest = (
-        tableCnt: number,
-        rowCnt: number,
-        truncated: boolean
-      ) => {
-        cy.log('confirm on 1hr')
-        cy.getByTestID('timerange-dropdown').within(() => {
-          cy.getByTestID('dropdown--button').should('exist')
-          cy.getByTestID('dropdown--button').clickAttached()
-        })
-        cy.getByTestID('dropdown-item-past1h').should('exist')
-        cy.getByTestID('dropdown-item-past1h').clickAttached()
-        cy.getByTestID('time-machine-submit-button').should('exist')
-        cy.getByTestID('time-machine-submit-button').clickAttached()
-
-        cy.log('run query')
-        cy.getByTestID('time-machine-submit-button').should('exist')
-        cy.getByTestID('time-machine-submit-button').clickAttached()
-        cy.wait('@query -1h')
-        cy.wait(1000) // bit more time for csv parsing
-
-        cy.log('table metadata displayed to user is correct')
-        if (truncated) {
-          cy.getByTestID('query-stat')
-            .should('contain', 'truncated')
-            .should('not.contain', 'rows')
-            .should('not.contain', 'tables')
-        } else {
-          cy.getByTestID('query-stat')
-            .should('not.contain', 'truncated')
-            .contains(`${tableCnt} tables`)
-          cy.getByTestID('query-stat').contains(`${rowCnt} rows`)
-        }
-
-        cy.log('will download complete csv data')
-        cy.intercept('POST', '/api/v2/query?*', req => {
-          req.redirect(route)
-        }).as('queryDownloadCSV')
-
-        // TODO: debug intermittent failures which occur 30% of the time
-        // in firefox only, in the CI only, and only when no data is returned
-        if (browser.name.toLowerCase() != 'chrome') {
-          return
-        }
-
-        cy.getByTestID('csv-download-button').should('not.be.disabled').click()
-
-        cy.wait('@queryDownloadCSV', {timeout: DELAY_FOR_FILE_DOWNLOAD})
-          .its('request', {timeout: DELAY_FOR_FILE_DOWNLOAD})
-          .then(req => {
-            cy.request(req)
-              .then(({body, headers}) => {
-                expect(headers).to.have.property(
-                  'content-type',
-                  'text/csv; charset=utf-8'
-                )
-                return Promise.resolve(body)
-              })
-              .then(csv => {
-                if (tableCnt == 0) {
-                  csv.trim() == ''
-                } else {
-                  validateCsv(csv, tableCnt)
-                }
-              })
-          })
-      }
-
-      beforeEach(() => {
-        cy.log('empty downloads directory')
-        cy.task('deleteDownloads', {dirPath: downloadsDirectory})
-      })
-
-      it('will return 0 tables and 0 rows, for an empty dataset', () => {
-        cy.isIoxOrg().then(isIox => {
-          // iox uses `${orgId}_${bucketId}` for a namespace_id
-          // And gives a namespace_id failure if no data is written yet.
-          // https://github.com/influxdata/monitor-ci/issues/402#issuecomment-1362368473
-          cy.skipOn(isIox)
-        })
-
-        cy.log('turn off composition sync')
-        cy.getByTestID('flux-sync--toggle').click()
-        cy.getByTestID('flux-sync--toggle').should('not.have.class', 'active')
-        cy.log('select empty dataset')
-        cy.getByTestID('flux-editor').monacoType(`{selectall}{enter}
-          from(bucket: "defbuck3") |> range(start: v.timeRangeStart, stop: v.timeRangeStop)
-        `)
-        cy.getByTestID('flux-editor').contains('defbuck3')
-
-        runTest(0, 0, false)
-      })
-
-      it('will return 1 table, for a dataset with only 1 table', () => {
-        cy.log('select dataset with 1 table')
-        cy.getByTestID('csv-download-button').should('be.disabled')
-        cy.getByTestID('flux-editor').monacoType(`{selectall}{enter}
-          from(bucket: "defbuck4") |> range(start: v.timeRangeStart, stop: v.timeRangeStop)
-          |> filter(fn: (r) => r._measurement == "ndbc_1table")
-        `)
-        cy.getByTestID('flux-editor').contains('defbuck4')
-
-        runTest(1, 1, false)
-      })
-
-      it('will return the complete dataset for smaller payloads', () => {
-        cy.log('select smaller dataset')
-        cy.getByTestID('flux-editor').monacoType(`{selectall}{enter}
-          from(bucket: "${bucketName}") |> range(start: v.timeRangeStart, stop: v.timeRangeStop)
-          |> filter(fn: (r) => r._measurement == "${measurement}")
-        `)
-        cy.getByTestID('flux-editor').contains(bucketName)
-
-        runTest(30, 30, false)
-      })
-
-      describe('for larger payloads', () => {
-        beforeEach(() => {
-          cy.setFeatureFlags({
-            newDataExplorer: true,
-            schemaComposition: true,
-            dataExplorerCsvLimit: 10000 as any,
-          })
-        })
-
-        it('will return a truncated dataset rows', () => {
-          cy.log('select larger dataset')
-          cy.getByTestID('csv-download-button').should('be.disabled')
-          cy.getByTestID('flux-editor').monacoType(`{selectall}{enter}
-            from(bucket: "defbuck4") |> range(start: v.timeRangeStart, stop: v.timeRangeStop)
-            |> filter(fn: (r) => r._measurement == "ndbc_big")
-          `)
-          cy.getByTestID('flux-editor').contains(
-            `|> filter(fn: (r) => r._measurement == "ndbc_big")`
-          )
-
-          runTest(numOfUniqueColumns, 2 * numOfUniqueColumns, true)
-        })
-      })
-    })
-  })
-
-  describe('Schema browser, without composition or saveAsScript', () => {
+  // Existing, old code. But turned off via feature flags.
+  describe.skip('Schema browser, without composition or saveAsScript', () => {
     beforeEach(() => {
       loginWithFlags({
         schemaComposition: false,
@@ -640,7 +411,7 @@ describe('Script Builder', () => {
         }).within(() => {
           cy.get('.composition-sync--on', {
             timeout: DELAY_FOR_LSP_SERVER_BOOTUP,
-          }).should('have.length', 3)
+          }).should('have.length.gte', 3)
         })
 
         cy.log('can still change composition')
@@ -668,18 +439,18 @@ describe('Script Builder', () => {
         confirmSchemaComposition()
 
         cy.log('sync toggles on and off, with matching styles')
-        cy.get('.composition-sync--on').should('have.length', 3)
+        cy.get('.composition-sync--on').should('have.length.gte', 3)
         cy.get('.composition-sync--off').should('have.length', 0)
         cy.getByTestID('flux-sync--toggle')
           .should('have.class', 'active')
           .click()
           .should('not.have.class', 'active')
         cy.get('.composition-sync--on').should('have.length', 0)
-        cy.get('.composition-sync--off').should('have.length', 3)
+        cy.get('.composition-sync--off').should('have.length.gte', 3)
         cy.getByTestID('flux-sync--toggle')
           .click()
           .should('have.class', 'active')
-        cy.get('.composition-sync--on').should('have.length', 3)
+        cy.get('.composition-sync--on').should('have.length.gte', 3)
         cy.get('.composition-sync--off').should('have.length', 0)
 
         cy.log('turn off flux sync')
